[SELC-6368] feat: Updated react-testing-library to latest version
