+++ tsconfig.json
@@ -24,5 +24,5 @@
     "resolveJsonModule": true
   },
   "include": ["src/**/*"],
-  "exclude": ["__mocks__", "dist", "node_modules", "**/__tests__/*", "Dangerfile.ts"]
+  "exclude": ["__mocks__", "dist", "node_modules", "**/__tests__/*", "Dangerfile.ts", "**/subProductStepPricingPlan/*"]
 }
